Load HRTF and ILD files from custom input (no longer hard coded) and prevent denormals for ARM64 platform
Builds successfully on Release x64 and ARM64 (unit tests need updating for other configurations.

+++ DiffractionPlugin/source/Spatialiser/Room.cpp
@@ -103,8 +103,8 @@ namespace UIE
 			{
 				std::vector<vec3> verticesB = b.GetVertices();
 
-				size_t numA = verticesA.size();
-				size_t numB = verticesB.size();
+				int numA = static_cast<int>(verticesA.size());
+				int numB = static_cast<int>(verticesB.size());
 
 				bool evenNum = numA % 2 == 0;
 				for (int i = 0; i < numA; i += 2)
